requirements.txt
langchain
wikipedia
langchain-community
langchain-anthropic
python-dotenv
pydantic
